splitting out gpconnect to stop circular imports
