+++ app/soap/responses.py
@@ -7,8 +7,8 @@ from datetime import datetime, timedelta
 import xmltodict
 from httpx import AsyncClient
 
+from ..gpconnect import gpconnect
 from ..redis_connect import redis_client
-from ..main import gpconnect
 
 
 def create_security():
@@ -228,7 +228,7 @@ async def iti_38_response(nhsno: int, queryid: str):
         # no cached ccda
         async with AsyncClient() as client:
             # r = await client.get(f"http://localhost:8000/gpconnect/{nhsno}")
-            #make internal call to gpconnect function
+            # make internal call to gpconnect function
             r = await gpconnect(nhsno)
             if r.status_code == 200:
                 logging.info(f"used internal call for {nhsno}")
